3D Printed Parts for Folding Quadcopter
Standoffs are expensive - so I printed my own. These are for M3 screws
(you may have to bore them out slightly with a 1/8” bit) and are 5/8”
high. Print with skirt for stability.

The second file is for skid type landing gear. These will take a lot of
force, so print with 3 or 4 perimeters and 25% infill.